+++ NEWS.md
@@ -1,3 +1,7 @@
+# fuj (development version)
+
+* additional local testing for os version
+
 # fuj 0.1.4
 
 * adds `%::%` and `%:::%` (which now aliases `%colons$`) or retrieving exported and non-exported values from namespaces [#31](https://github.com/jmbarbone/fuj/issues/31)