+++ NEWS.md
@@ -1,6 +1,7 @@
 # fuj (development version)
 
 * additional local testing for os version
+* `{covr}` dropped as a `suggest` 
 
 # fuj 0.1.4