+++ NEWS.md
@@ -1,5 +1,22 @@
 # fuj (development version)
 
+* `verbose()` has additional options for controlling the message output [#36](https://github.com/jmbarbone/fuj/issues/36)
+  * `.label` can be set to a string to prepend to the message (defaults to `"[verbose]"`)
+  * `.fill` can be set to repeat `.label` on each line of the message (defaults to `FALSE`)
+* `make_verbose()` is included to create a verbose function that will be triggered with a configured option, rather than the default `getOption("fuj.verbose", getOption("verbose"))` pattern.
+This can be used to define your own custom verbose function:
+
+```r
+library(fuj)
+options(fuj.verbose = FALSE)
+verbose("will not show")
+
+options(my.verbose = TRUE)
+my_verbose <- make_verbose("my.verbose")
+my_verbose("will show")
+#> [verbose] will show
+```
+
 * `%||%` is now set to only export in **R** versions < 4.4 [#35](https://github.com/jmbarbone/fuj/issues/35)
 * additional local testing for os version
 * `{covr}` dropped as a `suggest`